collect known nicknames independent of resource status
to avoid nickname conflicts if resources are unblocked
